requirements.txt
python-docx>=0.8.11
reportlab>=4.0.0
Pillow>=10.0.0
